+++ slackchannel2pdf/__init__.py
@@ -1,1 +1,6 @@
+import logging.config
+from . import settings
+
+logging.config.dictConfig(settings.DEFAULT_LOGGING)
+
 __version__ = "1.2.2"